refactor(agents): Migrate agent list to dedicated v2 route
Move agent list v2 implementation to dedicated route structure:
- Create new /agents-v2 route with separate layout and page components
- Add server action for agent creation in layout
- Simplify agent listing logic in page component
- Update redirect from legacy agents page

This reorganization improves code structure and separation of concerns while
preparing for v2 feature parity. The new route structure allows for better
scalability of the agents management interface.
