+++ app/(main)/agents/page.tsx
@@ -4,7 +4,6 @@ import { createAgent, getAgents } from "@/services/agents";
 import { CreateAgentButton } from "@/services/agents/components";
 import { redirect } from "next/navigation";
 import { Suspense } from "react";
-import { AgentListV2 } from "./v2";
 
 type AgentListProps = {
 	userId: string;
@@ -25,7 +24,7 @@ async function AgentList(props: AgentListProps) {
 export default async function AgentListPage() {
 	const enableV2 = await playgroundV2Flag();
 	if (enableV2) {
-		return <AgentListV2 />;
+		return redirect("/agents-v2");
 	}
 	const user = await getUser();
 	async function createAgentAction() {
